text_to_multi_speech/requirements.txt
piper-tts>=1.2.0
numpy>=1.21.0
sounddevice>=0.4.6
pyyaml>=6.0
requests>=2.28.0
langdetect>=1.0.9
pyttsx3
